fix サーバー側の検索機能をControllerに追加
