+++ public/index.html
@@ -20,7 +20,7 @@
             <!-- 検索フォーム -->
             <form id="searchForm">
                 <input type="text" id="searchInput" placeholder="ユーザー名またはメールアドレスを入力"> 
-                <button type="submit" id="searchButton">検索</button>
+                <button type="button" id="searchButton">検索</button>
             </form>
              <!-- 検索フォームここまで -->
             <a href="login.html">login</a>
