+++ server/models/userModel.js
@@ -72,20 +72,21 @@ const User = {
 
         // クエリを実行し、結果をコールバック関数に渡す
         db.query(query, [id], callback);
-    },
-
-    //検索条件でユーザーを検索するメソッド
-    Search: (query) => {
-      //検索条件に名前もしくはメールアドレスが一致したユーザーを取り出し、変数sqlに代入
-      const sql = `SELECT * FROM users WHERE name LIKE ? OR email LIKE ?`;
-      const values = [`%${query}%`, `%${query}%`];
-
-      //データベースから情報を取り出せなかったときはエラーを返し、取り出せた時は結果をjson形式で返す
-      db.query(sql, values, (err, results) => {
-        if (err) return res.status(500).send(err);
-        res.json(results);
-    });
     }
+
+//TODO:処理確認のためコントローラー側で確認。問題なければコメントアウトはずす。
+    // //検索条件でユーザーを検索するメソッド
+    // Search: (query) => {
+    //   //検索条件に名前もしくはメールアドレスが一致したユーザーを取り出し、変数sqlに代入
+    //   const sql = `SELECT * FROM users WHERE name LIKE ? OR email LIKE ?`;
+    //   const values = [`%${query}%`, `%${query}%`];
+
+    //   //データベースから情報を取り出せなかったときはエラーを返し、取り出せた時は結果をjson形式で返す
+    //   db.query(sql, values, (err, results) => {
+    //     if (err) return res.status(500).send(err);
+    //     res.json(results);
+    //  });
+    // }
 };
 
 // `User` オブジェクトをエクスポートし、他のファイルで使用できるようにする
